Fully Qualify BackendAddressPools in NetworkInterface Add api-version to Precheck Use Api-Version in convert to wssd object. Use to enforce name and other required params for V2 LBs. Broke out LB conversion into a new file
